requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.2
joblib>=1.1.0
rasterio>=1.2.10
matplotlib>=3.4.3
pillow>=8.3.2
sentinelsat>=1.1.1
pandas>=1.3.3
geopandas>=0.10.2
shapely>=1.8.0
requests>=2.26.0
python-dotenv>=0.19.0
tqdm>=4.62.3 